+++ angular-workspace/projects/component-library/src/lib/component-library.module.ts
@@ -1,9 +1,15 @@
+import { DialogModule } from "@angular/cdk/dialog";
+import { CommonModule } from "@angular/common";
 import { NgModule } from "@angular/core";
 
+import { DsoModalContainer } from "./components/modal-container/modal-container.component";
+import { DsoModalController } from "./controllers/modal/modal.controller";
 import { DIRECTIVES } from "./stencil-generated";
 
 @NgModule({
-  declarations: [...DIRECTIVES],
+  declarations: [...DIRECTIVES, DsoModalContainer],
+  imports: [CommonModule, DialogModule],
+  providers: [DsoModalController],
   exports: [...DIRECTIVES],
 })
 export class DsoToolkitModule {}