+++ angular-workspace/projects/component-library/src/lib/controllers/modal/index.ts
@@ -1,0 +1,2 @@
+export * from "./modal.controller";
+export * from "./modal-ref";