+++ angular-workspace/templates.ts
@@ -86,6 +86,7 @@ export interface Components {
   label: Label;
   listButton: ListButton;
   modal: Modal<StoryFnAngularReturnType>;
+  modalController: Modal<StoryFnAngularReturnType>;
   ozonContent: OzonContent;
   pagination: Pagination;
   progressBar: ProgressBar;