+++ packages/core/src/components/modal/modal.interfaces.ts
@@ -1,3 +1,5 @@
 export interface DsoModalCloseEvent {
   originalEvent?: MouseEvent;
 }
+
+export type ModalRole = "dialog" | "alert" | "alertDialog";