+++ packages/dso-toolkit/src/components/modal/modal.models.ts
@@ -13,3 +13,15 @@ export type ModalRole = "alert" | "dialog" | "alertdialog";
 export interface DsoModalCloseEvent {
   originalEvent?: MouseEvent;
 }
+
+export interface ModalContent<TemplateFnReturnType> {
+  title?: string;
+  body: TemplateFnReturnType;
+  footer?: TemplateFnReturnType;
+}
+
+export interface ModalOptions {
+  role?: ModalRole;
+  showCloseButton?: boolean;
+  initialFocus?: string;
+}