+++ packages/react/src/components/modal/modal.react-template.tsx
@@ -4,6 +4,7 @@ import * as React from "react";
 
 import { DsoModal } from "../../components";
 import { ComponentImplementation } from "../../templates";
+import { DsoModalPortal } from "../../react-components/modal-portal.component";
 
 export const reactModal: ComponentImplementation<Modal<JSX.Element>> = {
   component: "modal",
@@ -11,16 +12,20 @@ export const reactModal: ComponentImplementation<Modal<JSX.Element>> = {
   template: () =>
     function modalTemplate({ modalTitle, role, showCloseButton, initialFocus, body, footer, dsoClose }) {
       return (
-        <DsoModal
-          role={role}
-          modalTitle={modalTitle}
-          showCloseButton={showCloseButton}
-          initialFocus={initialFocus}
-          onDsoClose={(e: CustomEvent<DsoModalCloseEvent>) => dsoClose?.(e)}
-        >
-          <div slot="body">{body}</div>
-          {footer && <div slot="footer">{footer}</div>}
-        </DsoModal>
+        <DsoModalPortal>
+          <DsoModal
+            role={role}
+            modalTitle={modalTitle}
+            showCloseButton={showCloseButton ?? true}
+            initialFocus={initialFocus}
+            onDsoClose={(e: CustomEvent<DsoModalCloseEvent>) => dsoClose?.(e)}
+            // eslint-disable-next-line @typescript-eslint/no-explicit-any
+            ref={(el) => ((window as any)["modalRef"] = { framework: "react", ref: el })}
+          >
+            <div slot="body">{body}</div>
+            {footer && <div slot="footer">{footer}</div>}
+          </DsoModal>
+        </DsoModalPortal>
       );
     },
 };