+++ packages/react/src/index.ts
@@ -1,1 +1,3 @@
 export * from "./components";
+
+export * from "./react-components/modal-portal.component";