+++ packages/react/src/react-components/modal-portal.component.tsx
@@ -1,0 +1,6 @@
+import * as React from "react";
+import ReactDOM from "react-dom";
+
+export function DsoModalPortal({ children }: { children: React.ReactNode }) {
+  return <>{ReactDOM.createPortal(children, document.body)}</>;
+}