+++ storybook/src/components/modal/modal.core-template.ts
@@ -1,6 +1,6 @@
+import { DsoModalController } from "@dso-toolkit/core";
 import { Modal } from "dso-toolkit";
-import { html, TemplateResult } from "lit-html";
-import { ifDefined } from "lit-html/directives/if-defined.js";
+import { html, TemplateResult, render } from "lit-html";
 
 import { ComponentImplementation } from "../../templates";
 
@@ -9,17 +9,41 @@ export const coreModal: ComponentImplementation<Modal<TemplateResult>> = {
   implementation: "core",
   template: () =>
     function modalTemplate({ modalTitle, role, showCloseButton, initialFocus, body, footer, dsoClose }) {
-      return html`
-        <dso-modal
-          role=${role}
-          modal-title=${ifDefined(modalTitle)}
-          show-close-button=${ifDefined(showCloseButton)}
-          initial-focus=${ifDefined(initialFocus)}
-          @dsoClose=${dsoClose}
-        >
-          <div slot="body">${body}</div>
-          ${footer && html`<div slot="footer">${footer}</div>`}
-        </dso-modal>
-      `;
+      const open = () => {
+        /* eslint-disable @typescript-eslint/no-explicit-any */
+        if ((window as any).modalRef) {
+          (window as any).modalRef.close();
+          delete (window as any).modalRef;
+        }
+        /* eslint-enable @typescript-eslint/no-explicit-any */
+
+        const controller = new DsoModalController();
+
+        const bodyElement = document.createDocumentFragment();
+        const footerElement = document.createDocumentFragment();
+
+        render(body, bodyElement);
+        render(footer, footerElement);
+
+        const modalRef = controller.open(
+          {
+            title: modalTitle,
+            body: bodyElement,
+            footer: footerElement.childElementCount ? footerElement : undefined,
+          },
+          { role, showCloseButton, initialFocus }
+        );
+
+        if (dsoClose) {
+          modalRef.addEventListener("dsoClose", dsoClose as EventListenerOrEventListenerObject);
+        }
+
+        // eslint-disable-next-line @typescript-eslint/no-explicit-any
+        (window as any)["modalRef"] = modalRef;
+      };
+
+      open();
+
+      return html``;
     },
 };